requirements/develop.txt
-r production.txt
django==5.1.3
django-ninja==1.3.0